Commands refactoring and registry: - Moved commands to individual files - Added a simple registry
