+++ plusplusbot/command/gamestate_commands/gamestate_command.py
@@ -1,0 +1,7 @@
+from plusplusbot.command.commands import Command
+
+
+class GameStateCommand(Command):
+    def __init__(self, *args, **kwargs):
+        self.gamestate = kwargs.pop("gamestate")
+        super().__init__(*args, **kwargs)
