+++ plusplusbot/gamestate.py
@@ -1,15 +1,12 @@
-
-from plusplusbot.handlers import get_configuration_handler
-from plusplusbot.scorekeeper import InferredPlusPlusCommand
-from plusplusbot.wrappers import only_in_progress, admin_check
-from plusplusbot.commands import Command
+import json
+import logging
+import re
 from collections import defaultdict
 
-import logging
-import json
-import csv
-import re
-import io
+from plusplusbot.command.commands import Command
+from plusplusbot.handlers import get_configuration_handler
+from plusplusbot.command.scorekeeper_commands.inferred_plusplus_command import InferredPlusPlusCommand
+from plusplusbot.wrappers import only_in_progress, admin_check
 
 module_logger = logging.getLogger("PlusPlusBot.gamestate")
 
@@ -158,173 +155,3 @@ class GameState(object):
         self.state[channel]["winner"] = winner
         self.state[channel]["step"] = "waiting"
 
-    @property
-    def commands(self):
-        return [
-            SetAdmin,
-            RemoveAdmin,
-            NewGame,
-            SetEmojirade,
-        ]
-
-
-class GameStateCommand(Command):
-    def __init__(self, *args, **kwargs):
-        self.gamestate = kwargs.pop("gamestate")
-        super().__init__(*args, **kwargs)
-
-
-class SetAdmin(GameStateCommand):
-    pattern = "<@{me}> promote <@([0-9A-Z]+)>"
-    description = "Promotes a user to a game admin!"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["user"] = event["user"]
-        self.args["channel"] = event["channel"]
-
-        rendered_pattern = self.pattern.format(me=self.slack.bot_id)
-        self.args["admin"] = re.match(rendered_pattern, event["text"]).group(1)
-
-    @admin_check
-    def execute(self):
-        if self.gamestate.set_admin(self.args["channel"], self.args["admin"]):
-            yield (None, "<@{user}> has promoted <@{admin}> to a game admin :tada:".format(**self.args))
-        else:
-            yield (None, "<@{admin}> is already an admin :face_with_rolling_eyes:".format(**self.args))
-
-
-class RemoveAdmin(GameStateCommand):
-    pattern = "<@{me}> demote <@([0-9A-Z]+)>"
-    description = "Removes a user from the admins group"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["user"] = event["user"]
-        self.args["channel"] = event["channel"]
-
-        rendered_pattern = self.pattern.format(me=self.slack.bot_id)
-        self.args["admin"] = re.match(rendered_pattern, event["text"]).group(1)
-
-    @admin_check
-    def execute(self):
-        if self.gamestate.remove_admin(self.args["channel"], self.args["admin"]):
-            yield (None, "<@{user}> has demoted <@{admin}> to a pleb :cold_sweat:".format(**self.args))
-        else:
-            yield (None, "<@{admin}> isn't an admin :face_with_rolling_eyes:".format(**self.args))
-
-
-class NewGame(GameStateCommand):
-    pattern = "<@{me}> new game <@([0-9A-Z]+)> <@([0-9A-Z]+)>"
-    description = "Initiate a new game by setting the Old Winner and the Winner"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        rendered_pattern = self.pattern.format(me=self.slack.bot_id)
-
-        self.args["user"] = event["user"]
-        self.args["channel"] = event["channel"]
-        self.args["old_winner"] = re.match(rendered_pattern, event["text"]).group(1)
-        self.args["winner"] = re.match(rendered_pattern, event["text"]).group(2)
-
-    @admin_check
-    def execute(self):
-        if self.args["winner"] == self.args["old_winner"]:
-            yield (None, "Sorry, but the old and current winner cannot be the same person (<@{winner}>)...".format(**self.args))
-            raise StopIteration
-
-        self.gamestate.new_game(self.args["channel"], self.args["old_winner"], self.args["winner"])
-        yield (None, "<@{user}> has set the old winner to <@{old_winner}> and the winner to <@{winner}>".format(**self.args))
-        yield (None, "It's now <@{old_winner}>'s turn to provide <@{winner}> with the next 'rade!".format(**self.args))
-        yield (self.args["old_winner"], "You'll now need to send me the new 'rade for <@{winner}>".format(**self.args))
-        yield (self.args["old_winner"], "Please reply back in the format `emojirade Point Break` if `Point Break` was the new 'rade")
-
-
-class SetEmojirade(GameStateCommand):
-    pattern = "^emojirade (.+)$"
-    description = "Sets the current Emojirade to be guessed"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["user"] = event["user"]
-        self.args["emojirade"] = re.match(self.pattern, event["text"]).group(1)
-
-        # Figure out the channel to use
-        # TODO: Decide if we should be getting the user to enter the channel or not?
-        for channel in self.gamestate.state.keys():
-            if self.gamestate.state[channel]["step"] == "waiting":
-                self.args["channel"] = channel
-
-    @only_in_progress
-    def execute(self):
-        if self.args["user"] != self.gamestate.state[self.args["channel"]]["old_winner"]:
-            yield (None, "Err <@{user}> it's not your turn to provide the new 'rade :sweat:")
-            raise StopIteration
-
-        self.gamestate.set_emojirade(self.args["channel"], self.args["emojirade"])
-
-        winner = self.gamestate.state[self.args["channel"]]["winner"]
-
-        # DM the winner with the new rade
-        yield (winner, "Hey, <@{user}> made the 'rade `{emojirade}`, good luck!".format(**self.args))
-
-        # Let everyone else know
-        yield (self.args["channel"], ":mailbox: <@{user}> has sent the 'rade to <@{winner}>".format(**self.args, winner=winner))
-
-class InferredCorrectGuess(GameStateCommand):
-    pattern = None
-    description = "Takes the user that send the event as the winner, this is only ever fired internally"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["target_user"] = event["user"]
-        self.args["channel"] = event["channel"]
-        self.args["user"] = event["user"]
-
-    @only_in_progress
-    def execute(self):
-        self.gamestate.correct_guess(self.args["channel"], self.args["target_user"])
-
-        old_winner = self.gamestate.state[self.args["channel"]]["old_winner"]
-
-        yield (None, "<@{target_user}>++".format(**self.args))
-        yield (old_winner, "You'll now need to send me the new 'rade for <@{target_user}>".format(**self.args))
-        yield (old_winner, "Please reply back in the format `emojirade Point Break` if `Point Break` was the new 'rade")
-
-
-class CorrectGuess(GameStateCommand):
-    pattern = "<@([0-9A-Z]+)>[\s]*\+\+"
-    description = "Indicates the player has correctly guessed!"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["target_user"] = re.match(self.pattern, event["text"]).group(1)
-        self.args["channel"] = event["channel"]
-        self.args["user"] = event["user"]
-
-    @only_in_progress
-    def execute(self):
-        if self.args["user"] == self.args["target_user"]:
-            yield (None, "You're not allowed to award yourself the win >.>")
-            raise StopIteration
-
-        if self.args["user"] != self.gamestate.state[self.args["channel"]]["old_winner"]:
-            yield (None, "You're not the old winner, stop awarding other people the win >.>")
-            raise StopIteration
-
-        self.gamestate.correct_guess(self.args["channel"], self.args["target_user"])
-        yield (None, "<@{target_user}>++".format(**self.args))
-        yield (self.args["user"], "You'll now need to send me the new 'rade for <@{target_user}>".format(**self.args))
-        yield (self.args["user"], "Please reply back in the format `emojirade Point Break` if `Point Break` was the new 'rade")
