+++ plusplusbot/scorekeeper.py
@@ -1,13 +1,12 @@
-
-from plusplusbot.handlers import get_configuration_handler
-from plusplusbot.commands import Command
-from plusplusbot.wrappers import only_in_progress, admin_check
+import csv
+import io
+import logging
+import re
 from collections import defaultdict
 
-import logging
-import csv
-import re
-import io
+from plusplusbot.command.commands import Command
+from plusplusbot.handlers import get_configuration_handler
+from plusplusbot.wrappers import only_in_progress, admin_check
 
 module_logger = logging.getLogger("PlusPlusBot.scorekeeper")
 
@@ -82,196 +81,3 @@ class ScoreKeeper(object):
 
     def save(self):
         self.config.save(self.scoreboard)
-
-    @property
-    def commands(self):
-        return [
-            PlusPlusCommand,
-            MinusMinusCommand,
-            SetCommand,
-            LeaderboardCommand,
-            HistoryCommand,
-        ]
-
-
-class ScoreKeeperCommand(Command):
-    def __init__(self, *args, **kwargs):
-        self.scorekeeper = kwargs.pop("scorekeeper")
-        self.gamestate = kwargs.pop("gamestate")
-        super().__init__(*args, **kwargs)
-
-
-class InferredPlusPlusCommand(ScoreKeeperCommand):
-    pattern = None
-    description = "Increments the users score"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["target_user"] = event["user"]
-        self.args["channel"] = event["channel"]
-        self.args["user"] = event["user"]
-
-    @only_in_progress
-    def execute(self):
-        target_user = self.args["target_user"]
-
-        self.logger.debug("Incrementing user's score: {}".format(target_user))
-        self.scorekeeper.plusplus(target_user)
-
-        score = self.scorekeeper.scoreboard[target_user]
-
-        message = "Congrats <@{0}>, you're now at {1} point{2}"
-        yield (None, message.format(target_user, score, "s" if score > 1 else ""))
-
-    def __str__(self):
-        return "InferredPlusPlusCommand"
-
-
-class PlusPlusCommand(ScoreKeeperCommand):
-    pattern = "<@([0-9A-Z]+)>[\s]*\+\+"
-    description = "Increment the users score"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["target_user"] = re.match(self.pattern, event["text"]).group(1)
-        self.args["channel"] = event["channel"]
-        self.args["user"] = event["user"]
-
-    @only_in_progress
-    def execute(self):
-        target_user = self.args["target_user"]
-
-        if self.args["user"] == target_user:
-            yield (None, ":thinking_face: you're not allowed to award points to yourself...")
-            raise StopIteration
-
-        if self.slack.is_bot(target_user):
-            yield (None, ":thinking_face: robots aren't allowed to play Emojirades!")
-            raise StopIteration
-
-        self.logger.debug("Incrementing user's score: {}".format(target_user))
-        self.scorekeeper.plusplus(target_user)
-
-        score = self.scorekeeper.scoreboard[target_user]
-
-        message = "Congrats <@{0}>, you're now at {1} point{2}"
-        yield (None, message.format(target_user, score, "s" if score > 1 else ""))
-
-    def __str__(self):
-        return "PlusPlusCommand"
-
-
-class SetCommand(ScoreKeeperCommand):
-    pattern = "<@([0-9A-Z]+)> set (-?[0-9]+)"
-    description = "Manually set the users score"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["user"] = event["user"]
-        self.args["channel"] = event["channel"]
-
-        matches = re.match(self.pattern, event["text"])
-        self.args["target_user"] = matches.group(1)
-        self.args["new_score"] = matches.group(2)
-
-    @admin_check
-    def execute(self):
-        target_user = self.args["target_user"]
-        new_score = int(self.args["new_score"])
-
-        if self.args["user"] == target_user:
-            yield ":thinking_face: you can't do that to yourself"
-            raise StopIteration
-
-        if self.slack.is_bot(target_user):
-            yield ":thinking_face: robots aren't allowed to play Emojirades"
-            raise StopIteration
-
-        self.logger.debug("Setting {} score to: {}".format(target_user, new_score))
-        self.scorekeeper.overwrite(target_user, new_score)
-
-        message = "<@{0}> manually set to {1} point{2}"
-        yield (None, message.format(target_user, new_score, "s" if new_score > 1 else ""))
-
-    def __str__(self):
-        return "SetCommand"
-
-
-class MinusMinusCommand(ScoreKeeperCommand):
-    pattern = "<@([0-9A-Z]+)> --"
-    description = "Decrement the users score"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def prepare_args(self, event):
-        self.args["target_user"] = re.match(self.pattern, event["text"]).group(1)
-        self.args["channel"] = event["channel"]
-        self.args["user"] = event["user"]
-
-    @admin_check
-    def execute(self):
-        target_user = self.args["target_user"]
-
-        if self.args["user"] == target_user:
-            yield ":thinking_face: you're not allowed to deduct points from yourself..."
-            raise StopIteration
-
-        if self.slack.is_bot(target_user):
-            yield ":thinking_face: robots aren't allowed to play Emojirades!"
-            raise StopIteration
-
-        self.logger.debug("Decrementing user's score: {}".format(target_user))
-        self.scorekeeper.minusminus(target_user)
-
-        score = self.scorekeeper.scoreboard[target_user]
-
-        message = "Oops <@{0}>, you're now at {1} point{2}"
-        yield (None, message.format(target_user, score, "s" if score > 1 else ""))
-
-    def __str__(self):
-        return "MinusMinusCommand"
-
-
-class LeaderboardCommand(ScoreKeeperCommand):
-    pattern = "<@{me}> leaderboard"
-    description = "Shows all the users scores"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def execute(self):
-        leaderboard = self.scorekeeper.leaderboard()
-
-        self.logger.debug("Printing leaderboard: {0}".format(leaderboard))
-
-        yield (None, "\n".join(["{0}. <@{1}> [{2} point{3}]".format(index + 1, name, score, "s" if score > 1 else "")
-                               for index, (name, score) in enumerate(leaderboard)]))
-
-    def __str__(self):
-        return "LeaderboardCommand"
-
-
-class HistoryCommand(ScoreKeeperCommand):
-    pattern = "<@{me}> history"
-    description = "Shows the latest few actions performed"
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def execute(self):
-        history = self.scorekeeper.history()
-
-        self.logger.debug("Printing history: {0}".format(history))
-
-        yield (None, "\n".join(["{0}. <@{1}> > '{2}'".format(index + 1, name, action)
-                               for index, (name, action) in enumerate(history)]))
-
-    def __str__(self):
-        return "HistoryCommand"
